requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-3-jet==1.0.8
django-ckeditor==6.0.0
django-filebrowser==3.13.2
django-grappelli==2.14.3
django-import-export==2.5.0
django-js-asset==1.2.2
django-tinymce==3.2.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.5
PyYAML==5.4
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
